select multiple creates separate state for each variant
